Reworked Gold a little bit
Modify the current implementation of the custom gold, so that it doesn't
keep dota gold at 0, but keeps the dota gold between 0 and 50000. This
way, you can make all the shop purchases work as normal.

Please note that in this kind of system purchasing items that cost more
than 50000 is impossible if they can't be combined from other items...
If they do combine, it is still not possible to buy it directly, but you
must by the "parts" first until the "gold to complete" is lower than
50000.

Of course, you could make the limit higher. Let's say 80000 or 90000 for
example... It pretty much depends on how much "dota bounty" you get from
regular dota events (player kills for example).