Added Siege Mode item
This is going to need a pass-over by the balance team, as I kept the
values from the keyvalues that were already there - which, due to the
fact that there was no number for splash damage, means that it deals
100% splash damage. Have fun exploiting this before they get to it!

Will also need another pass-over code-wise later on, both to implement
active icons ( trying to do so ahead of time would just result in game
crashes ) and to fix the fact that you can shove it into your backpack
and keep it active, like with Stoneskin Armor. The latter of which I
struggled to fix, but OnInventoryContentsChanged just seems to be dead
with this repo.