Dagon
Dagon did far too much damage previously. I have significantly nerfed
the damage but also added in a massive int bonus. This int bonus adds up
to 15% spell amp, which should help compensate for the lack of raw
damage compared to before. Mostly an experiment, I expect people to not
really analyze this change properly and assume the item was nerfed to
shit, while in some situations it's probably as strong as before.