Reworked MKB
Previously MKBs bash was incredibly powerful but otherwise the item was
merely okay. Often the only times you did any real damage was when the
minibash happened and otherwise you couldn't quite hurt people. This
change reduces the power of the minibash but reintroduces that power on
the regular damage of the item, keeping the overall dps similar.  I
compensated for the loss of pure damage by doubling the amount when
changing it to physical damage.