Balance (#978)
* Dagon

Dagon did far too much damage previously. I have significantly nerfed
the damage but also added in a massive int bonus. This int bonus adds up
to 15% spell amp, which should help compensate for the lack of raw
damage compared to before. Mostly an experiment, I expect people to not
really analyze this change properly and assume the item was nerfed to
shit, while in some situations it's probably as strong as before.

* Assault Cuirass

Adjust AC to be more reasonable.

* Radiance

Radiance was a bit on the stronger end, especially level 2 was
overwhelming if gotten early. I nerfed it across the board although
mostly the early game power, late is a little weaker as well but not as
much.

* Bristleback

Bristle is currently the strongest hero in the game. I have nerfed the
early max damage on Quill Spray. This hardly impacts Bristles fighting
ability since to deal the maximum damage he needs to get 5 stacks on the
target. Against Creeps however this makes his early farming rotation
about 2-3 Quill Sprays slower. On top of that I have slightly increased
the mana cost on Quill Spray.

I also introduced a cooldown on Bristleback and slightly nerfed his
other abilities. Not sure if the cooldown will just work like this but
we'll see.

* Heart Nerf

Small nerf, mostly meant to make lvl 2 heart a bit easier to get through

* Lucience

Small nerf

Taking babysteps in lucience balance atm to slowly approach a state
where the item is fine. I don't think it's op currently, just a bit on
the strong side.

* Desolator

A mild buff on Desolators minus armor. The item currently seems worse
than Daedalus and MKB but considering we also just nerfed armor with the
removal of SC I want to be careful with buffing this too much.

* Reworked MKB

Previously MKBs bash was incredibly powerful but otherwise the item was
merely okay. Often the only times you did any real damage was when the
minibash happened and otherwise you couldn't quite hurt people. This
change reduces the power of the minibash but reintroduces that power on
the regular damage of the item, keeping the overall dps similar.  I
compensated for the loss of pure damage by doubling the amount when
changing it to physical damage.

* Adjusted Assault a bit more

It is now completely in line with Shivas Guard