Covfefe
Pretty massive buff. The item currently sees absolutely no play. I think
a big reason why is that it gave worse stats than the items it combined
from, so upgrading literally made you weaker. It now gives more stats
than the items it combines from, but not as much as the basic upgrade
paths would give you (which is how I want combination items to work).

May be overbuffed, especially the damage. But seeing as currently
literally nobody ever uses it, I rather have it be on the stronger side
so I can see how it plays out and then nerf it down again.