Balance (#1112)
* Shiva and BF

Shiva adjusted to be weaker. It's attack speed slow is now somewhat in
line with regular dota in comarison to AC.
BF mild buff.

* Blink Dagger

I want to see if we can make Blink Dagger be viable while not adding any
stats to it.

* BKB

Mild BKB buff, the items effect is good but feels bad to buy since it
provides almost no stats. Now it provides a little more in terms of
stats.

* Silencer

Hero is dumb. Nerfed all his stuff, expect him to still be dumb.

* Phantom Lancer

Reduced the amount of illusions on ultimate mostly for lag purposes.
Still scales on %, which should be good enough.

* Elder Titan

He still did too much damage. His scaling is now in line with other
heroes. Before it was ridiculously high for no reason.

* Bubble Orb

Reduced Radius

* Shiva Adjusted

Mild rescaling