Refactor calls to math.random to use Dota API random functions instead
If I did this right, most of this should be largely inconsequential. The main exceptions being the Charger tower locations, which looks like it was using math.random in such a way that the location vector could only be 1 of 9 locations, and ward_invisiblity which was using math.random in a way it's not really meant to be used.