Balance (#1140)
* Siege Mode

Part of Siege Mode design was always to reduce AS. So we have introduced
that. The item before this nerf was super strong, should be
significantly weaker now. Will require testing if it's too weak.

* Drow Ranger

Her aura often granted upwards of 350 dmg to her team, which is far
higher than it should be.

* OD

Q is a % based ability. Those shouldn't scale as much as others since
they scale naturally.

* Simple Boss Fury Swipes

Guy was far too hard. Deals half damage now.

* Covfefe

Removed Desolator from Covfefe upgrade path. This is a massive buff so I
reduced it's damage to avoid making it overpowered. The item was super
expensive, now it isn't anymore. I'll be baffled if it still doesn't see
play.

* Lucience

Previously a max level Lucience didn't require an upgrade core 1. It now
does.

* Force Staff

Reduced Cooldown and made it show up in the tooltip.

* Lotus Orb

Made it a hopefully viable alternative to AC and Shiva

* Urn of Shadows

Trying to turn what is usually only an early game support item into a
viable late game tool. Mostly an experiment.

* Fixes

* More fixes