Slight buff to arc warden
This is a slight buff to arc warden in hopes that the hero will become more viable.
Make a viable arcanes strat where you can farm the creeps and get the gold or send the illu off to get assist gold else where
Plus the thing is then alive longer. obviously a buff.

If this is still not enough we can lower cooldown, but I don't like the idea of making the duration any longer than this

Also this is an attempted nerf to silencer as arc is one of his worst matchups in vanilla.