Heart
Heart was significantly out of line in terms of scaling and power level
as other items. It is now of a lower power level, making more builds
viable in terms of dealing with it and also allowing different tank
items to be relevant.
It also lost it's early game "free tank" abilities. As you upgrade
you'll regain the ability to tank bosses easily.