Balance (#1258)
* Heart

Heart was significantly out of line in terms of scaling and power level
as other items. It is now of a lower power level, making more builds
viable in terms of dealing with it and also allowing different tank
items to be relevant.
It also lost it's early game "free tank" abilities. As you upgrade
you'll regain the ability to tank bosses easily.

* MKB

Part of the big change to damage items in general. Reducing the overall
DPS of rightclickers by about 10-15%.

* Daedalus

Nerfed as part of the big DPS nerf

* Battlefury

Decreased cleave distance to make getting the big cleaves a bit more
difficult.

* Desolator

Nerfed as part of the big DPS nerf

* Reflex buffs

Made levels 1 and 2 a bit stronger to make them more viable.