Mekansm
After some testing I found the item to be weak in all aspects but not
significantly. So I buffed most parts which should now work quite well.
The decently high armor combined with the regeneration aura is a very
effective tool to keep supports alive. The AOE heal may need a bit more
of a buff but honestly the rest of the item seems quite strong as is, I
don't want to overdo it.