New Underlord Ultimate
We have replaced Underlords ultimate since it was a useless ability in
OAA. The new ability opens up a portal in 2 locations that allow any
friends and enemies to travel through them every 3 seconds. This is the
old Underlords ultimate. We hope you enjoy this new ability, balance is
still in the works and will need testing first.