Balance (#1310)
* Boss Magic Resistance

All Bosses now have -25% magic resistance and are more vulnerable to
disables. This should make casters more viable Boss killers. Will need
testing to see if this makes killing bosses too easy and if so we should
adjust their health.

* Helm of the Dominator

Dominator now has proper stats. Before it was too weak to be considered
viable, now it may be too strong. It's the first iteration of the item,
I expect to have to change it.

* Heavens Halberd

If the active doesn't get more powerful then the mana cost shouldn't
increase.

* Mekansm

After some testing I found the item to be weak in all aspects but not
significantly. So I buffed most parts which should now work quite well.
The decently high armor combined with the regeneration aura is a very
effective tool to keep supports alive. The AOE heal may need a bit more
of a buff but honestly the rest of the item seems quite strong as is, I
don't want to overdo it.

* Butterfly

Mild buff. I am being very careful with this item because it holds the
potential to break the game if it's too strong. Overbuffing it even a
tiny bit would invalidate tons of item builds, so we are taking baby
steps here.

* Greater GGs

Buffed. Currently TP boots are better in almost every regard. This
should make it a bit more even.

* New Underlord Ultimate

We have replaced Underlords ultimate since it was a useless ability in
OAA. The new ability opens up a portal in 2 locations that allow any
friends and enemies to travel through them every 3 seconds. This is the
old Underlords ultimate. We hope you enjoy this new ability, balance is
still in the works and will need testing first.

* Underlord Fix