Update Highway to Hell
Precaches the portal particle so it properly shows up, as well as adds a
"Close Highway" subspell to prematurely close the portals.

As a result, there's now a status buff on Underlord that shows him the
duration of all his portals, and units are interrupted ( Nether Swap
style ) when they go through a portal.