Refactor Aghanim's, make undroppable on Undying (#1704)
* Refactor Aghanim's Scepter to remove modifiers to disable passive Talents

* Make modifier_item_aghanims_talents not RemoveOnDeath

This is so that Talents aren't disable and re-enabled on death and respawn

* Make modifier_intrinsic_multiplexer assert existence of its modifiers on respawn

This allows for modifiers that are RemoveOnDeath to be used with modifier_intrinsic_multiplexer, since the multiplexer doesn't get recreated on death and respawn.

* Make Aghanim's above level 1 permanent on Undying

This is necessary as there is no way to disable his reincarnation Talent. Removing the modifier allows refreshing of the cooldown by dropping and picking up Agh's.

* Remove SetCanBeUsedOutOfInventory in Aghanim's

Thought it might be related to allowing item in backpack. Testing shows that it isn't, and that non-droppable items automatically can't be put in backpack.