Balance (#1747)
* Temple Guardian nerf

Halfed HP since their previous total combined HP was twice that of a
normal tier 4. Reduced stun durations to make the fight less annoying.
Nerfed the hammer throw damage

* Undying Nerf

Reduced Zombie spawn rate to prevent trapping with Naga + Undying.

* Rapier Nerf

Nerfed damage by 500. Increased purchase requirements to 3 Upgrade Core
4.

* Heart Transplant

Nerfed and reworked to be a normal item that doesn't require a million
cores.

* Lycan Boss

Reduced HP to be a bit fairer as a Tier 2.

* Ogre Nerf

Reduced HP to compensate for the added difficulty of the Seers.

* Temple Guardians

* 20k