Fix non-working uses of UseResources method (#1771)
* Fix non-working uses of UseResources method

Apparently Valve went and broke this method for most types of abilities. Not really sure what cases still work.

- Fix Time Lock and Bash of the Deep not having a cooldown
- Fix Glaives of Wisdom not using mana

* Revert "Fix non-working uses of UseResources method"

This reverts commit b06b014c938f0109254793df376e54c59ca30d42.

* Add ShouldUseResources true for various items that call UseResources

Apparently Valve added this ShouldUseResources thing, or changed it to default to false in a bunch of cases. Why in the world it would need to default to false, I may never know.