Implement custom aura system, rework Fool's Gold (#1796)
* Add base classes for custom aura implementation

This implementation handles similar auras overlapping much better than vanilla Dota, as it keeps track of all item/ability sources and returns the highest level or oldest, valid source from calls to GetAbility on the aura effect modifier. Notably, this fixes issues with upgrading items, or spawning illusions causing auras to stop working. It also allows for multiple similar auras from a single unit to stack.

* Fix caster of aura modifiers not updating when items are passed between units

* Make modifier_creep_bounty use the custom aura system

Also made it non-stacking, and have a stickiness of 10 seconds, to prevent drop/backpack-to-farm abuse.

* Remove Greater Arcanes and Greater Tranquils from aura_item_upgrade list

The custom aura system should negate the need for this, and the 10 second stickiness of Fool's Gold would render the workaround ineffective anyway.