Changed Respawn Time
Respawn Time was changed from set 5s to set 10s. The biggest change in
this patch. Will require feedback on this to see if it's a good change
or not. I just really want to try it.

Also reduced the time you have to wait for the game to end if a team
abandons.