Fixed some bugs with the Viper Strike recode (#2298)
* Change Viper silence talent

Now applies to Viper Strike instead of Nethertoxin. Currently uses the
same rules as the rest of the ability, i.e. pieces BKB, but this is
changable and this seems fair anyway given that it's moving from 8
seconds in an AoE to 5 seconds on a single target

* if you say so travis

* Fixed some bugs with the Viper Strike recode

namely the fact that I forgot to check if the projectile actually hit a
valid target whoops

* meh