Lifestealer Rage and Feast Nerf  (#2521)
* Lifestealer Nerf

* Update lifestealer.txt

Changed the bonus strength to all stats, as it is a more suitable replacement to his old talent.

* Update lifestealer.txt

* Update lifestealer.txt

Left the file in for if people want to make future changes to lifestealer, but removed the changes I made.

* Update life_stealer_rage.txt

Nerfed Rage duration, capping it at 6 seconds. This measn the talent doesn't have to be changed anymore, and lifestealer can't get past 8 seconds of spell immunity. Reduced the attack speed slightly at level 6.

* Update life_stealer_rage.txt

* Lifestealer nerfs (+feast)