Visage Buffs and Shard fix (#3133)
* Visage Buffs (looking for ideas and feedback)

Small changes to Q to ease my eyes of the pain and too allow oaa values to be a bit higher, as it should be.

Gravekeeper cloak heavy nerfs reverted since the max dmg cap is added to ability.

Buffs to familiar to cosign with all the other player units that have been buffed. Familiars will now do bit more dmg and be a bit more tanky but still should be killable in all respective manners. Like to see what people think

* fixes

darkonis will yell at me for this xd

* Update visage_gravekeepers_cloak.lua

* Update visage_summon_familiars_oaa.txt

* Update visage_summon_familiars_oaa.txt

* Update visage_summon_familiars_oaa.txt

* Update visage_summon_familiars_oaa.txt

* Update visage_summon_familiars_oaa.txt

* Update visage.txt

* Update npc_abilities_custom.txt

* Create npc_abilities_custom.txt

* Update visage_gravekeepers_cloak_oaa.txt

* Update visage_gravekeepers_cloak_oaa.txt

* Draft version of shard functionality

* Fixed Gravekeeper's Cloak not being disabled with break.
* Trying to make Visage shard functional.

* Visage Shard now works

* Visage already has Stone Form ability but hidden. Adding a new Stone Form ability to Visage creates a crash.

* Tooltip updates

Co-authored-by: Darko V <darko1290@gmail.com>